requirements.txt
# Discord bot
discord.py>=2.3.0
python-dotenv>=1.0.0

# HTTP and web scraping
aiohttp>=3.9.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
fake-useragent>=1.4.0

# Search
ddgs>=6.0.0

# Excel and charts
openpyxl>=3.1.0
xlsxwriter>=3.1.0
matplotlib>=3.8.0
pandas>=2.1.0

# AI/LLM via Ollama
ollama>=0.1.0

# Web GUI
flask>=3.0.0
flask-socketio>=5.3.0
flask-cors>=4.0.0

# MCP (Model Context Protocol)
mcp>=1.0.0

# Scheduling / automation
apscheduler>=3.10.0

# PDF parsing (for articles/patents)
pypdf2>=3.0.0
pdfplumber>=0.10.0

# Config
pyyaml>=6.0.0

# Async support
asyncio-throttle>=1.0.0

# Data processing
numpy>=1.26.0
